Move PurchaseMetaOwner component into its own tsx file
